try to fix test for arm & riscv 32 bit

+++ tests/src/utl/meta/src/get_common_type.cpp
@@ -20,7 +20,8 @@ TEST_CASE("get_common_type" * doctest::test_suite("utl"))
         NMTOOLS_STATIC_CHECK_IS_SAME( result_t, expect_t );
     }
     {
-        using arg_t = utl::tuple<int,size_t>;
+        // using arg_t = utl::tuple<int,size_t>;
+        using arg_t = utl::tuple<int,unsigned long>;
         using result_t = meta::get_common_type_t<arg_t>;
         using expect_t = long int;
         NMTOOLS_STATIC_CHECK_IS_SAME( result_t, expect_t );
